+++ src/client/main.tsx
@@ -1,5 +1,26 @@
 import * as React from "react";
-import { unstable_createRoot } from "react-dom";
+import { Root, unstable_createRoot } from "react-dom";
 import App from "./App";
 
-unstable_createRoot(document.getElementById("main")!).render(<App />);
+const root: Root = unstable_createRoot(document.getElementById("main")!);
+root.render(<App />);
+
+declare var module: {
+    hot?: any;
+}
+
+if (module.hot) {
+    module.hot.accept("./App", (err: unknown) => {
+        try {
+            const NextApp = require("./App").default;
+            root.render(<NextApp />);
+            return;
+        } catch (e) {
+            console.error(e);
+            // to be reloaded
+        }
+        console.error(err);
+        console.log("reloading self");
+        window.location.reload();
+    });
+}