+++ src/webpack.shared.config.js
@@ -1,7 +1,8 @@
 const webpack = require("webpack");
 const path = require("path");
 
-const mode = process.env.NODE_ENV || "production";
+const isDevelopment = process.env.NODE_ENV !== "production";
+const mode = isDevelopment ? "development" : "production";
 
 console.log(`webpack operating under mode \`${mode}\``);
 
@@ -31,15 +32,6 @@ module.exports = {
 
   module: {
     rules: [
-      {
-        test: /\.ts(x?)$/,
-        exclude: /node_modules/,
-        use: [
-            {
-                loader: "ts-loader"
-            }
-        ]
-      },
       {
         test: /\.s[ac]ss$/i,
         use: [